move zone actions to village panel
